+++ lib/IMAP/ImapMessageFetcher.php
@@ -42,8 +42,8 @@ use OCA\Mail\Exception\ServiceException;
 use OCA\Mail\IMAP\Charset\Converter;
 use OCA\Mail\Model\IMAPMessage;
 use OCA\Mail\Service\Html;
-use OCA\Mail\Service\SmimeService;
 use OCA\Mail\Service\PhishingDetection\PhishingDetectionService;
+use OCA\Mail\Service\SmimeService;
 use OCP\AppFramework\Db\DoesNotExistException;
 use function str_starts_with;
 use function strtolower;
