+++ lib/PhishingDetectionList.php
@@ -34,36 +34,36 @@ use ReturnTypeWillChange;
  */
 class PhishingDetectionList implements JsonSerializable {
 
-/** @var PhishingDetectionResult[] */
-   private $checks; 
+	/** @var PhishingDetectionResult[] */
+	private $checks;
 
-   private bool $warning = false;
+	private bool $warning = false;
 
-    public function __construct(array $checks = []) {
-         $this->checks = $checks;
-    }
+	public function __construct(array $checks = []) {
+		$this->checks = $checks;
+	}
 
-    public function addCheck(PhishingDetectionResult $check) {
-        $this->checks[] = $check;
-    }
-    private function isWarning() {
-        foreach ($this->checks as $check) {
-            if (in_array($check->getType(), [PhishingDetectionResult::DATE_CHECK, PhishingDetectionResult::LINK_CHECK, PhishingDetectionResult::CUSTOM_EMAIL_CHECK, PhishingDetectionResult::CONTACTS_CHECK]) && $check->isPhishing()) {
-                return true;
-            }
-        }
-        return false;
-    }
+	public function addCheck(PhishingDetectionResult $check) {
+		$this->checks[] = $check;
+	}
+	private function isWarning() {
+		foreach ($this->checks as $check) {
+			if (in_array($check->getType(), [PhishingDetectionResult::DATE_CHECK, PhishingDetectionResult::LINK_CHECK, PhishingDetectionResult::CUSTOM_EMAIL_CHECK, PhishingDetectionResult::CONTACTS_CHECK]) && $check->isPhishing()) {
+				return true;
+			}
+		}
+		return false;
+	}
 
-    #[ReturnTypeWillChange]
+	#[ReturnTypeWillChange]
 	public function jsonSerialize() {
 		$result = array_map(static function (PhishingDetectionResult $check) {
 			return $check->jsonSerialize();
 		}, $this->checks);
-        return [
-            'checks' => $result,
-            'warning' => $this->isWarning(),
-        ];
+		return [
+			'checks' => $result,
+			'warning' => $this->isWarning(),
+		];
 	}
 
 }
